BindingAdaptersのファイル分割
View毎に分割する。
